product-url-crawler/scraper-trigger/requirements.txt
functions-framework==3.3.0
google-cloud-firestore==2.7.0
google-cloud-storage==2.7.0
requests==2.31.0